Refactor separate button component into nested button component
