+++ components/generic_desk/__init__.py
@@ -6,12 +6,10 @@ from esphome.const import CONF_ID
 
 DEPENDENCIES = ["uart"]
 AUTO_LOAD = [
-    "desk_memory_button",
     "desk_height_sensor",
     "desk_is_moving_sensor",
     "sensor",
     "binary_sensor",
-    "button",
 ]
 MULTI_CONF = True
